initial phase a routing
routed connections for phase A not including SPI. spi left as stubs that can be routed to connector and to the other ADCs.
Power and PWRGND routing is done on layer 1,2,4 and can be discussed further in persons. For now I have VDC on layer 1 and 4. PWRGND on layer 1 and 4. SW and PHA on layer 1 and 2. layer 3 and 5 are gnd. Layer 1 and 6 are signal.